gpu_metrics: update get_average_gfx_power type to Option<u32>

+++ amdgpu/gpu_metrics.rs
@@ -92,7 +92,7 @@ impl MetricsInfo for GpuMetrics {
     impl_metrics!(get_average_socket_power, Option<u32>);
     impl_metrics!(get_average_cpu_power, Option<u16>);
     impl_metrics!(get_average_soc_power, Option<u16>);
-    impl_metrics!(get_average_gfx_power, Option<u16>);
+    impl_metrics!(get_average_gfx_power, Option<u32>);
     impl_metrics!(get_average_core_power, Option<Vec<u16>>);
     impl_metrics!(get_average_gfxclk_frequency, Option<u16>);
     impl_metrics!(get_average_socclk_frequency, Option<u16>);
